learn

+++ libft/ft_strlen.c
@@ -6,15 +6,15 @@
 /*   By: yajallal < yajallal@student.1337.ma >      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/09/29 16:47:34 by yajallal          #+#    #+#             */
-/*   Updated: 2022/10/17 00:11:02 by yajallal         ###   ########.fr       */
+/*   Updated: 2023/02/04 14:42:16 by yajallal         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-size_t	ft_strlen(const char *s)
+int	ft_strlen(const char *s)
 {
-	size_t	i;
+	int	i;
 
 	i = 0;
 	while (s[i] != '\0')
